fix(core): merge FilterInput and CombinedFilterInput into one module
The two classes are mutually recursive and lived in two files that imported
each other. FilterInput referenced CombinedFilterInput EAGERLY — in a decorator
argument (`type: CombinedFilterInput`) and in the `design:type` metadata that
emitDecoratorMetadata emits for `combinedFilter?: CombinedFilterInput`. Both are
evaluated at class-definition time, i.e. while the module is still initializing,
so on the cycle they read a class binding still in its temporal dead zone.

This was not latent. It crashed, today, in shipped code:

  require('dist/core/common/inputs/combined-filter.input.js')
  -> ReferenceError: Cannot access 'CombinedFilterInput' before initialization

It stayed hidden only because entering through the package barrel pulls
filter.input in first via other importers (filter.args, filter.helper). A
vendor-mode deep import, a unit test importing the input directly, or a
reordering of src/index.ts would all have hit it.

A lazy thunk does NOT fix this: `type: () => CombinedFilterInput` still leaves
the eager design:type metadata behind, and SWC's `typeof` guard does not protect
the member expression it compiles to. The only fix is to remove the import edge,
so both classes now live in filter.input.ts with CombinedFilterInput declared
FIRST (that ordering is load-bearing). combined-filter.input.ts stays as a
re-export, so every existing import path still resolves.
