+++ src/core/common/inputs/combined-filter.input.ts
@@ -1,57 +1,10 @@
-import { InputType } from '@nestjs/graphql';
-
-import { Restricted } from '../decorators/restricted.decorator';
-import { UnifiedField } from '../decorators/unified-field.decorator';
-import { LogicalOperatorEnum } from '../enums/logical-operator.enum';
-import { RoleEnum } from '../enums/role.enum';
-import { maps } from '../helpers/model.helper';
-import { CoreInput } from './core-input.input';
-import { FilterInput } from './filter.input';
-
-@InputType({
-  description: 'Combination of multiple filters via logical operator',
-})
-@Restricted(RoleEnum.S_EVERYONE)
-export class CombinedFilterInput extends CoreInput {
-  /**
-   * Logical Operator to combine filters
-   */
-  @UnifiedField({
-    description: 'Logical Operator to combine filters',
-    enum: LogicalOperatorEnum,
-    roles: RoleEnum.S_EVERYONE,
-  })
-  logicalOperator: LogicalOperatorEnum = undefined;
-
-  /**
-   * Filters to combine via logical operator
-   */
-  @UnifiedField({
-    description: 'Filters to combine via logical operator',
-    isOptional: true,
-    roles: RoleEnum.S_EVERYONE,
-    type: () => FilterInput,
-  })
-  filters: FilterInput[] = undefined;
-
-  // ===================================================================================================================
-  // Methods
-  // ===================================================================================================================
-
-  /**
-   * Mapping for Subtypes
-   */
-  override map(
-    data: Partial<this> | Record<string, any>,
-    options: {
-      cloneDeep?: boolean;
-      funcAllowed?: boolean;
-      mapId?: boolean;
-    } = {},
-  ): this {
-    super.map(data, options);
-    this.filters = maps(data.filters, FilterInput, options.cloneDeep);
-    Object.keys(this).forEach((key) => this[key] === undefined && delete this[key]);
-    return this;
-  }
-}
+/**
+ * `CombinedFilterInput` is declared in `./filter.input` — the two classes are mutually recursive
+ * and must live in one module, or the cycle between them crashes SWC-compiled builds with
+ * `ReferenceError: Cannot access 'CombinedFilterInput' before initialization`. See the docblock in
+ * `./filter.input` for the full analysis.
+ *
+ * This file remains only so existing deep imports of `.../inputs/combined-filter.input` keep
+ * resolving. Do NOT move the class back here.
+ */
+export { CombinedFilterInput } from './filter.input';
